Rewrote the queue code because I was suspicous of the old code.
In the end, I have no proof the old code was broken.

The new code is simpler to follow and uses only basic locking.

I would like to put in a condition variable so that one can do
a get-with-timeout -- and then the main thread can avoid the
usleep(250000) step.