Added queue_get_wtimeout()
Waits for a queue element to become available up to the given timeout, otherwise returns NULL.

After eventually figuring out that the main skydiveorbust thread is not called from where I thought it was,
I have no immediate use for this feature, but it is tested and it works as expected.   It's superior to
a loop that sleeps & periodically polls the queue because latency is minimized.