Added GUI 'roundlist' page for selecting team/round among the list available from the current video input.
It's crude, but it works - and avoids the need to type at a keyboard
slowly and make mistakes.

This special mode does not interfere with prior operation if the name
of the video file does not match the special pattern.

If files are chopped up as one-flight-per-file, the user never needs
to browse the options - the team#/round# is already correct.