Workflow fix.
Un-did an old feature -- if the current video file is named to declare
only a single flight, the team/round "buttons" were disabled since
there's only one valid choice -- however, I had not considered that a
submission will clear those fields.  If the judge immediately realizes
an error after submitting and wants to re-judge, he should be able to
do so without having to re-open the file -- so re-enable the buttons
when #flights is > 0, rather than > 1.