Fixed bug with videolist selection operation.
Things were happening in different threads in an unpredictable sequence.

When the sdob page got re-opened, the auto-set team#/round# was
getting wiped out.

So, I made loading a video a queued operation to defer it to the right time.

No changes in _videolist.c or main.c needed for this fix -- the calls
from both places were affected equally.