Fixed a crash bug that has been around for a while.
If no file is selected in the main page and the SDOB button is
pressed...

OR

If no file is selected in the _videolist page and the Change Video
button is pressed...

it was crashing by dereferencing an array at the -1 index.  This is
now fixed.  One can now enter SDOB without selecting a file, and the
ChangeVideo button press is ignored if no file is selected.  If one
wishes to return to the sdob page without changing the video, the
Cancel button is available.